Move API functions to api.php
